+++ lib/rfi-beeline.d.ts
@@ -19,29 +19,32 @@ export declare class HoneycombBeelineFactory {
     make(requestId: string, beelineImplementation?: any): RFIBeeline;
 }
 declare abstract class Beeline {
-    withSpan<T>(metadataContext: object, fn: (span: HoneycombSpan) => T, rollupKey?: string): T;
-    withAsyncSpan<T>(this: Beeline, metadataContext: object, fn: () => T): Promise<T>;
-    withTrace<T>(metadataContext: object, fn: () => T, withTraceId?: string, withParentSpanId?: string, withDataset?: string): T;
-    startTrace(metadataContext: object, traceId?: string, parentSpanId?: string, dataset?: string): HoneycombSpan;
+    withSpan<T>(metadataContext: Record<string, unknown>, fn: (span: HoneycombSpan) => T, rollupKey?: string): T;
+    withAsyncSpan<T>(this: Beeline, metadataContext: Record<string, unknown>, fn: () => T): Promise<T>;
+    withTrace<T>(metadataContext: Record<string, unknown>, fn: () => T, withTraceId?: string, withParentSpanId?: string, withDataset?: string): T;
+    startTrace(metadataContext: Record<string, unknown>, traceId?: string, parentSpanId?: string, dataset?: string): HoneycombSpan;
     finishTrace(span: HoneycombSpan): void;
-    startSpan(metadataContext: object, spanId?: string, parentId?: string): HoneycombSpan;
+    startSpan(metadataContext: Record<string, unknown>, spanId?: string, parentId?: string): HoneycombSpan;
     finishSpan(span: HoneycombSpan, rollup?: string): void;
-    startAsyncSpan<T>(metadataContext: object, fn: (span: HoneycombSpan) => T): T;
+    startAsyncSpan<T>(metadataContext: Record<string, unknown>, fn: (span: HoneycombSpan) => T): T;
     bindFunctionToTrace<T>(fn: () => T): T;
     runWithoutTrace<T>(fn: () => T): T;
-    addContext(context: object): void;
-    removeContext(context: object): void;
+    addContext(context: Record<string, unknown>): void;
+    removeContext(context: Record<string, unknown>): void;
     marshalTraceContext(context: HoneycombSpan): string;
     unmarshalTraceContext(context_string: string): HoneycombSpan;
     getTraceContext(): HoneycombSpan;
 }
+/**
+ * @deprecated
+ */
 export declare class RFIBeeline extends Beeline {
     requestId: string;
     private _beelineImplementation;
     constructor(requestId: string, beelineImplementation?: any);
     get beeline(): any;
-    withSpan<T>(metadataContext: object, fn: (span: HoneycombSpan) => T): T;
-    withAsyncSpan(this: RFIBeeline, spanData: any, spanFn: Function): Promise<any>;
+    withSpan<T>(metadataContext: Record<string, unknown>, fn: (span: HoneycombSpan) => T): T;
+    withAsyncSpan(this: RFIBeeline, spanData: Record<string, unknown>, spanFn: Function): Promise<any>;
     bindFunctionToTrace<T>(fn: () => T): T;
 }
 export interface IHoneycombBeelineFactory {