+++ lib/with-span.decorator.d.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata';
-export declare function WithSpan(context?: object): MethodDecorator;
+export declare function WithSpan(context?: Record<string, unknown>): MethodDecorator;
 /**
  *
  * @param span_metadata?: {}  metadata to add to the Honeycomb trace context.
@@ -9,4 +9,4 @@ export declare function WithSpan(context?: object): MethodDecorator;
  *                    method: `${wrappedMethod.name}`
  *                  }
  */
-export declare function AddToTrace(span_metadata?: object): MethodDecorator;
+export declare function AddToTrace(span_metadata?: Record<string, unknown>): MethodDecorator;